dbt-airflow/requirements.txt
pydantic
python-dotenv
pydantic-settings
apache-airflow-providers-trino